添加详情页Detail.vue
