+++ package.json
@@ -14,11 +14,11 @@
     "vue": "^2.6.11",
     "vue-meta": "^2.3.3",
     "vue-router": "^3.1.6",
-    "vuetify": "^2.2.26",
-    "vuex": "^3.3.0"
+    "vuetify": "^2.2.28",
+    "vuex": "^3.4.0"
   },
   "devDependencies": {
-    "@mdi/font": "^5.1.45",
+    "@mdi/font": "^5.2.45",
     "@vue/cli-plugin-babel": "^4.3.1",
     "@vue/cli-plugin-eslint": "^4.3.1",
     "@vue/cli-plugin-pwa": "^4.3.1",
@@ -34,10 +34,10 @@
     "eslint-plugin-standard": "^4.0.1",
     "eslint-plugin-vue": "^6.2.2",
     "highlight.js": "^9.18.1",
-    "lint-staged": "^10.2.0",
+    "lint-staged": "^10.2.2",
     "lodash": "^4.17.15",
     "markdown-it": "^10.0.0",
-    "markdown-it-anchor": "^5.2.7",
+    "markdown-it-anchor": "^5.3.0",
     "markdown-it-container": "^2.0.0",
     "markdown-it-emoji": "^1.4.0",
     "markdown-it-table-of-contents": "^0.4.4",
