+++ src/views/photo/Index.vue
@@ -15,11 +15,8 @@ export default {
   mixins: [
     LoadSections([
       'hero-alt',
-      'about-our-product',
       'theme-features',
-      'affiliates',
       'newsletter',
-      'pro-features',
       'info'
     ])
   ],
